# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yajja delivery & social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (for invite search)
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Vendors ───────────────────────────────────────────────────────────────────
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      parameters:
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/Category"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Register vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorBody"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{vendorId}:
    get:
      operationId: getVendor
      tags: [vendors]
      parameters:
        - $ref: "#/components/parameters/vendorId"
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    put:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - $ref: "#/components/parameters/vendorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{vendorId}/stats:
    get:
      operationId: getVendorStats
      tags: [vendors]
      summary: Vendor dashboard stats
      parameters:
        - $ref: "#/components/parameters/vendorId"
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorStats"

  # ── Products ──────────────────────────────────────────────────────────────────
...
  /riders/register:
    post:
      operationId: registerRider
      tags: [riders]
      summary: Register as a rider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRiderBody"
      responses:
        "201":
          description: Rider registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderProfile"
  /riders/me:
    get:
      operationId: getRiderProfile
      tags: [riders]
      summary: Get rider profile
      responses:
        "200":
          description: Rider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderProfile"
    put:
      operationId: updateRiderLocation
      tags: [riders]
      summary: Update rider location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderProfile"
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
    vendorId:
      name: vendorId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Category:
      type: string
      enum: [food, liquor, pharmacy, household]
...
    UserRole:
      type: string
      enum: [customer, vendor, rider, admin]

    # ── Auth ────────────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: string
        avatarUrl:
          type: string
        address:
          type: string
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]

    # ── Vendor ───────────────────────────────────────────────────────────────────
    Vendor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        description:
          type: string
        imageUrl:
          type: string
        address:
          type: string
        rating:
          type: number
        deliveryTime:
          type: string
        minOrder:
          type: number
        isOpen:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, name, category, isOpen, createdAt]
    CreateVendorBody:
      type: object
      properties:
        name:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        description:
          type: string
        imageUrl:
          type: string
        address:
          type: string
        deliveryTime:
          type: string
        minOrder:
          type: number
      required: [name, category]
    UpdateVendorBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        address:
          type: string
        deliveryTime:
          type: string
        minOrder:
          type: number
        isOpen:
          type: boolean
    VendorStats:
      type: object
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        todayRevenue:
          type: number
        weekRevenue:
          type: number
        totalRevenue:
          type: number
        avgRating:
          type: number
        ordersThisWeek:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              count:
                type: integer
              revenue:
                type: number
            required: [day, count, revenue]
      required: [totalOrders, pendingOrders, todayRevenue, weekRevenue, totalRevenue, avgRating, ordersThisWeek]

    # ── Product ───────────────────────────────────────────────────────────────────
...
    RiderProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        vehicleType:
          type: string
        licensePlate:
          type: string
        currentLat:
          type: number
        currentLng:
          type: number
        isAvailable:
          type: boolean
        totalDeliveries:
          type: integer
        rating:
          type: number
        createdAt:
          type: string
      required: [id, userId, isAvailable, totalDeliveries, createdAt]
    RegisterRiderBody:
      type: object
      properties:
        vehicleType:
          type: string
        licensePlate:
          type: string
      required: [vehicleType, licensePlate]
    UpdateLocationBody:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        isAvailable:
          type: boolean
      required: [lat, lng]